# backend_laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TelEat API Documentation",
        "description": "Dokumentasi API untuk aplikasi pemesanan makanan TelEat (Laravel & React)",
        "contact": {
            "email": "admin@teleat.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/pesanans": {
            "get": {
                "tags": [
                    "Pesanan"
                ],
                "summary": "Daftar pesanan",
                "description": "Mengambil daftar riwayat pesanan. Pelanggan hanya melihat pesanan miliknya, Merchant hanya melihat pesanan ke tokonya, dan Admin melihat semua.",
                "operationId": "10a3468ed30d0e7f352e84d91ced8ff6",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pesanan"
                ],
                "summary": "Buat pesanan baru",
                "description": "Hanya pelanggan yang dapat membuat pesanan. Mengurangi stok menu secara otomatis dan membuat transaksi dengan metode pembayaran CASH atau MIDTRANS.",
                "operationId": "6dd8d55376bf4211203eed581ee5a510",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "merchant_id",
                                    "items"
                                ],
                                "properties": {
                                    "merchant_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Pedas sedang ya",
                                        "nullable": true
                                    },
                                    "nomor_meja": {
                                        "type": "string",
                                        "example": "5",
                                        "nullable": true
                                    },
                                    "tipe_pemesanan": {
                                        "type": "string",
                                        "enum": [
                                            "DINE_IN",
                                            "TAKE_AWAY"
                                        ],
                                        "example": "DINE_IN"
                                    },
                                    "metode_bayar": {
                                        "type": "string",
                                        "enum": [
                                            "CASH",
                                            "MIDTRANS"
                                        ],
                                        "example": "CASH"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "menu_id",
                                                "jumlah"
                                            ],
                                            "properties": {
                                                "menu_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "jumlah": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pesanan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stok tidak mencukupi"
                    },
                    "403": {
                        "description": "Hanya pelanggan yang bisa memesan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pesanans/{id}/status": {
            "put": {
                "tags": [
                    "Pesanan"
                ],
                "summary": "Update status pesanan",
                "description": "Hanya merchant pemilik pesanan yang dapat memperbarui status pesanan. Mengubah status transaksi menjadi LUNAS atau BATAL sesuai status baru.",
                "operationId": "d5cbbbdf10903c7906511cb799a1f3ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "PENDING",
                                            "DIPROSES",
                                            "SELESAI",
                                            "BATAL"
                                        ],
                                        "example": "DIPROSES"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diubah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Status tidak dapat diubah lagi"
                    },
                    "403": {
                        "description": "Tidak diizinkan"
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pesanans/{id}/batal": {
            "put": {
                "tags": [
                    "Pesanan"
                ],
                "summary": "Batalkan pesanan",
                "description": "Hanya pelanggan pemilik pesanan yang dapat membatalkan pesanan jika belum SELESAI atau BATAL. Mengkembalikan stok menu yang dipesan.",
                "operationId": "581e66e323e21c33cfd36b0cb40bfc29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pesanan berhasil dibatalkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pesanan berhasil dibatalkan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Pesanan sudah selesai atau batal"
                    },
                    "403": {
                        "description": "Tidak diizinkan"
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pesanans/{id}": {
            "get": {
                "tags": [
                    "Pesanan"
                ],
                "summary": "Detail pesanan",
                "description": "Mengambil rincian detail satu pesanan beserta relasinya (pelanggan, merchant, detail pesanan, menu, dan transaksi).",
                "operationId": "fc4900e18b46ff883ed5eb05562c0fa9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token Sanctum (auth_token) hasil login untuk mengakses endpoint yang dilindungi",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}